refactor(pdf): update trial dog PDF template handling and introduce new templates
- Modified the `getTrialDogPdfTemplateFileName` function to return null for unsupported templates.
- Enhanced error handling in PDF rendering for unknown rule windows.
- Refactored PDF rule set definitions into a new `registry.ts` file, consolidating template paths and statuses.
- Added new PDF templates for 2011-2023 and updated the 2005-2011 template.
- Removed outdated templates and ensured compatibility with the new structure.

Files affected:
- Updated logic in `trial-dog-pdf.ts` and `trial-dog-pdf.test.ts`.
- Introduced `registry.ts` for rule set management.
- Added new PDF files and removed obsolete ones.
