+++ apps/web/lib/public/beagle/trials/pdf/__tests__/trial-dog-pdf.test.ts
@@ -11,7 +11,7 @@ import {
 describe("renderTrialDogPdf", () => {
   it("renders pdf bytes when lisatiedot sections are omitted", async () => {
     const bytes = await renderTrialDogPdf({
-      trialRuleWindowId: null,
+      trialRuleWindowId: "trw_post_20110801",
       registrationNo: "FI12345/21",
       dogName: null,
       dogSex: "MALE",
@@ -137,13 +137,13 @@ describe("renderTrialDogPdf", () => {
   });
 
   it.each([
-    ["trw_pre_20020801", "ajok-koirakohtainen-poytakirja.pdf"],
-    ["trw_range_2002_2005", "ajok-koirakohtainen-poytakirja.pdf"],
-    ["trw_range_2005_2011", "ajok-koirakohtainen-poytakirja-2005-2011.pdf"],
-    ["trw_post_20110801", "ajok-koirakohtainen-poytakirja.pdf"],
-    ["trw_post_20230801", "ajok-koirakohtainen-poytakirja.pdf"],
-    [null, "ajok-koirakohtainen-poytakirja.pdf"],
-    ["unknown-window", "ajok-koirakohtainen-poytakirja.pdf"],
+    ["trw_pre_20020801", null],
+    ["trw_range_2002_2005", null],
+    ["trw_range_2005_2011", "ajok-poytakirja-2005-2011.pdf"],
+    ["trw_post_20110801", "ajok-poytakirja-2011-2023.pdf"],
+    ["trw_post_20230801", null],
+    [null, null],
+    ["unknown-window", null],
   ])(
     "selects the PDF template for rule window %s",
     (ruleWindowId, expected) => {
@@ -188,67 +188,64 @@ describe("renderTrialDogPdf", () => {
   });
 
   it("does not fall back to another timeline for unknown rule windows", async () => {
-    const bytes = await renderTrialDogPdf({
-      trialRuleWindowId: "unknown-window",
-      registrationNo: "SHOULD-NOT-RENDER-UNKNOWN",
-      dogName: "SHOULD-NOT-RENDER-DOG",
-      dogSex: "MALE",
-      sireName: null,
-      sireRegistrationNo: null,
-      damName: null,
-      damRegistrationNo: null,
-      omistaja: null,
-      omistajanKotikunta: null,
-      kennelpiiri: null,
-      kennelpiirinro: null,
-      koekunta: null,
-      koemaasto: null,
-      koepaiva: new Date("2025-09-07T00:00:00.000Z"),
-      jarjestaja: null,
-      era1Alkoi: null,
-      era2Alkoi: null,
-      hakuMin1: null,
-      hakuMin2: null,
-      ajoMin1: null,
-      ajoMin2: null,
-      hyvaksytytAjominuutit: null,
-      ajoajanPisteet: null,
-      hakuEra1: null,
-      hakuEra2: null,
-      hakuKeskiarvo: null,
-      haukkuEra1: null,
-      haukkuEra2: null,
-      haukkuKeskiarvo: null,
-      hakuloysyysTappioEra1: null,
-      hakuloysyysTappioEra2: null,
-      hakuloysyysTappioYhteensa: null,
-      ajoloysyysTappioEra1: null,
-      ajoloysyysTappioEra2: null,
-      ajoloysyysTappioYhteensa: null,
-      tappiopisteetYhteensa: null,
-      ajotaitoEra1: 4,
-      ajotaitoEra2: 2,
-      ajotaitoKeskiarvo: 3,
-      ansiopisteetYhteensa: 0,
-      loppupisteet: 0,
-      paljasMaaTaiLumi: null,
-      luopui: false,
-      suljettu: false,
-      keskeytetty: false,
-      koetyyppi: "NORMAL",
-      sijoitus: null,
-      koiriaLuokassa: null,
-      Palkinto: "1",
-      huomautusTeksti: null,
-      ryhmatuomariNimi: null,
-      palkintotuomariNimi: null,
-      ylituomariNumeroSnapshot: null,
-      ylituomariNimiSnapshot: null,
-    });
-
-    const rawPdf = Buffer.from(bytes).toString("latin1");
-
-    expect(rawPdf).not.toContain("SHOULD-NOT-RENDER-UNKNOWN");
-    expect(rawPdf).not.toContain("SHOULD-NOT-RENDER-DOG");
+    await expect(
+      renderTrialDogPdf({
+        trialRuleWindowId: "unknown-window",
+        registrationNo: "SHOULD-NOT-RENDER-UNKNOWN",
+        dogName: "SHOULD-NOT-RENDER-DOG",
+        dogSex: "MALE",
+        sireName: null,
+        sireRegistrationNo: null,
+        damName: null,
+        damRegistrationNo: null,
+        omistaja: null,
+        omistajanKotikunta: null,
+        kennelpiiri: null,
+        kennelpiirinro: null,
+        koekunta: null,
+        koemaasto: null,
+        koepaiva: new Date("2025-09-07T00:00:00.000Z"),
+        jarjestaja: null,
+        era1Alkoi: null,
+        era2Alkoi: null,
+        hakuMin1: null,
+        hakuMin2: null,
+        ajoMin1: null,
+        ajoMin2: null,
+        hyvaksytytAjominuutit: null,
+        ajoajanPisteet: null,
+        hakuEra1: null,
+        hakuEra2: null,
+        hakuKeskiarvo: null,
+        haukkuEra1: null,
+        haukkuEra2: null,
+        haukkuKeskiarvo: null,
+        hakuloysyysTappioEra1: null,
+        hakuloysyysTappioEra2: null,
+        hakuloysyysTappioYhteensa: null,
+        ajoloysyysTappioEra1: null,
+        ajoloysyysTappioEra2: null,
+        ajoloysyysTappioYhteensa: null,
+        tappiopisteetYhteensa: null,
+        ajotaitoEra1: 4,
+        ajotaitoEra2: 2,
+        ajotaitoKeskiarvo: 3,
+        ansiopisteetYhteensa: 0,
+        loppupisteet: 0,
+        paljasMaaTaiLumi: null,
+        luopui: false,
+        suljettu: false,
+        keskeytetty: false,
+        koetyyppi: "NORMAL",
+        sijoitus: null,
+        koiriaLuokassa: null,
+        Palkinto: "1",
+        huomautusTeksti: null,
+        ryhmatuomariNimi: null,
+        palkintotuomariNimi: null,
+        ylituomariNumeroSnapshot: null,
+        ylituomariNimiSnapshot: null,
+      }),
+    ).rejects.toThrow("PDF rule set unsupported has no template.");
   });
 });
